+++ 김지훈/jpashop/src/main/java/jpabook/jpashop/repository/MemberOldRepository.java
@@ -1,0 +1,35 @@
+package jpabook.jpashop.repository;
+
+import java.util.List;
+
+import org.springframework.stereotype.Repository;
+
+import jakarta.persistence.EntityManager;
+import jpabook.jpashop.domain.Member;
+import lombok.RequiredArgsConstructor;
+
+@Repository
+@RequiredArgsConstructor
+public class MemberOldRepository {
+
+	private final EntityManager em;
+
+	public void save(Member member) {
+		em.persist(member);
+	}
+
+	public Member findOne(Long id) {
+		return em.find(Member.class, id);
+	}
+
+	public List<Member> findAll() {
+		return em.createQuery("select m from Member m", Member.class)
+			.getResultList();
+	}
+
+	public List<Member> findByName(String name) {
+		return em.createQuery("select m from Member m where m.name = :name", Member.class)
+			.setParameter("name", name)
+			.getResultList();
+	}
+}
