+++ 김지훈/jpashop/src/main/java/jpabook/jpashop/repository/MemberRepository.java
@@ -2,34 +2,11 @@ package jpabook.jpashop.repository;
 
 import java.util.List;
 
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import jakarta.persistence.EntityManager;
 import jpabook.jpashop.domain.Member;
-import lombok.RequiredArgsConstructor;
 
-@Repository
-@RequiredArgsConstructor
-public class MemberRepository {
+public interface MemberRepository extends JpaRepository<Member, Long> {
 
-	private final EntityManager em;
-
-	public void save(Member member) {
-		em.persist(member);
-	}
-
-	public Member findOne(Long id) {
-		return em.find(Member.class, id);
-	}
-
-	public List<Member> findAll() {
-		return em.createQuery("select m from Member m", Member.class)
-			.getResultList();
-	}
-
-	public List<Member> findByName(String name) {
-		return em.createQuery("select m from Member m where m.name = :name", Member.class)
-			.setParameter("name", name)
-			.getResultList();
-	}
+	List<Member> findByName(String name);
 }
