+++ 김지훈/jpashop/src/main/java/jpabook/jpashop/service/OrderService.java
@@ -31,7 +31,7 @@ public class OrderService {
 	@Transactional
 	public Long order(Long memberId, Long itemId, int count) {
 		// 엔티티 조회
-		Member member = memberRepository.findOne(memberId);
+		Member member = memberRepository.findById(memberId).get();
 		Item item = itemRepository.findOne(itemId);
 
 		// 배송정보 생성
